use generic user image instead of vanity user image
Before this commit: To add a user container, we build a custom
image and data container for the user, such that the custom image
has a user account with vanity name. For example:

* base image: jumanjiman/booga
* vanity image: someuser
* data container: someuser-data
* runtime container: someuser-run

This has drawbacks:

* We must build a custom image for each user.
* We have to track that custom image.
* We have to coordinate upgrade with the user if the underlying
  base image needs update.

After this commit: Every user gets a generic runtime container
with a custom data container.

* base image: jumanjiman/booga
* data container: someuser-data
* runtime container: someuser-run

We can throw away the runtime container at any time.
Upon startup, the new runtime container is the latest version
of the base image.